Creating EstimateContainer test and fixing some broken tests
